Ensure html_escape marks result as html_safe
When `<%= h("{"id":"example"}") %>`is evaluated, `h` references
`ERB::Util#html_escape_with_escaping_angular_expressions_erb` which
calls out first to our custom `Escaper`, and the result is then passed
to the original `#html_escape` method that was saved off as
`#html_escape_without_escaping_angular_expressions_erb` when creating a
decorated alias. The original method then performs the escaping, marking
the result as `html_safe`.

In this case, we expect
`#html_escape_without_escaping_angular_expressions_erb` points to
`ERB::Util#html_escape` which performs the escaping and marks the result
as `html_safe`. However, the alias points to `Haml::Helpers#html_escape`
which does NOT mark the result as `html_safe`. As such, we encounter the
following problem:

`<div attribute="<%= h('{"key":"value"}') %>"></div>`

1. Any Ruby snippets in the template are executed:
`h('{"key":"value"}')` => `'{&quot;key&quot;:&quot;value&quot;}'`

2. Rails auto-escapes all attributes by default
`'{&quot;key&quot;:&quot;value&quot;}'` =>
`'{&amp;quot;key&amp;quot;:&amp;quot;value&amp;quot;}'`

Because the result of (1) is not marked as `html_safe`, instances of
`"&"` are re-escaped during the auto-escaping step, resulting in an
invalid attribute value when rendered by the client.

Our change ensures that whatever we return from
`ERB::Util#html_escape_with_escaping_angular_expressions_erb` and
`Haml::Helpers#html_escape_with_escaping_angular_expressions_haml` is
marked as `html_safe` to prevent double-escaping.

Follow-Up:
We've tested the 4 combinations of using `h` and `html_escape` in both
`.html.haml` and `.html.erb` files to better understand the code paths
that are executed in this monkey-patched world we live in.

The results are interesting:

* From within an ERB file:
  * html_escape
    * Points to Haml::Helpers::XssMods#html_escape_with_haml_xss
    (NOTE: This method returns an `html_safe` string once escaped—see)
      * Which then calls to
      Haml::Helpers#html_escape_with_escaping_angular_expressions_haml
        * Which then calls to the original Haml::Helpers#html_escape
  * h
    * Points to
    ERB::Util#html_escape_with_escaping_angular_expression_erb
      * Which then calls to the original ERB::Util#html_escape
      (NOTE: This method returns an `html_safe` string once escaped)
* From within a HAML file:
  * html_escape
    * Points to Haml::Helpers::XssMods#html_escape_with_haml_xss
    (NOTE: This method returns an `html_safe` string once escaped—see)
      * Which then calls to
      Haml::Helpers#html_escape_with_escaping_angular_expressions_haml
        * Which then calls to the original Haml::Helpers#html_escape
    * h
      * Points to
      ERB::Util#html_escape_with_escaping_angular_expression_erb
        * Which then calls to the original ERB::Util#html_escape
        (NOTE: This method returns an `html_safe` string once escaped)

Note that in all cases, the string ends up `html_escaped` by an existing
function, but also note that we get some unexpected behavior--Haml
calls are made from within `.html.erb` files, and ERB calls are made
within `.html.haml` files. The results are difficult to predict, and
likely depend on load-order and monkey-patching order, something that
is difficult to control.

As such, we choose to mark the results of our aliased methods as
`html_safe` as an added layer of protection in case one of the calls
in the chain ever unknowingly points to a call to `html_escape` which
does NOT mark the result as `html_safe`. But have no fear, `html_safe`
is an idempotent operation, so we won't harm anything that's already
been marked as such.